rename : domain 패키지 생성 후 meber와 voucher domain 이동
